reorganizing to make this easier to scale and maintain
